refactor: 디렉토리 구조 Atomic Design Pattern으로 변경
