+++ components/atoms/ExternalLink.tsx
@@ -5,7 +5,7 @@ import { Platform } from 'react-native';
 
 type Props = Omit<ComponentProps<typeof Link>, 'href'> & { href: string };
 
-export function ExternalLink({ href, ...rest }: Props) {
+function ExternalLink({ href, ...rest }: Props) {
   return (
     <Link
       target="_blank"
@@ -22,3 +22,5 @@ export function ExternalLink({ href, ...rest }: Props) {
     />
   );
 }
+
+export default ExternalLink;
