+++ components/atoms/Icon/index.tsx
@@ -1,15 +1,16 @@
-import { type ComponentProps } from 'react';
 import { colors } from '@/theme';
 import HomeIcon from '@/assets/images/HomeIcon';
 import AuthIcon from '@/assets/images/auth.svg';
 import MyIcon from '@/assets/images/my.svg';
 import MyOutlineIcon from '@/assets/images/my-outline.svg';
 import AuthOutlineIcon from '@/assets/images/auth-outline.svg';
+import { SvgProps } from 'react-native-svg';
 
-interface Icons {
-  name: 'home' | 'home-outline' | 'auth' | 'auth-outline' | 'mypage' | 'mypage-outline';
+export type IconTypes = 'home' | 'home-outline' | 'auth' | 'auth-outline' | 'mypage' | 'mypage-outline';
+interface IconProps extends SvgProps {
+  name: IconTypes;
 }
-function TabBarIcon({ name, ...rest }: ComponentProps<typeof HomeIcon> & Icons) {
+function Icon({ name, ...rest }: IconProps) {
   switch (name) {
     case 'home':
       return <HomeIcon fill={colors.primary} stroke={colors.primary} {...rest} />;
@@ -24,7 +25,6 @@ function TabBarIcon({ name, ...rest }: ComponentProps<typeof HomeIcon> & Icons) 
     case 'mypage-outline':
       return <MyOutlineIcon {...rest} />;
   }
-  return null;
 }
 
-export default TabBarIcon;
+export default Icon;
